+++ src/main/java/com/ashish/graphqldemo/graphql/input/UpdateUserInput.java
@@ -16,7 +16,7 @@ public class UpdateUserInput {
 
     private String email;
 
-    private Integer monthlySalary;
+    private Float monthlySalary;
 
-    private Integer monthlyExpense;
+    private Float monthlyExpense;
 }
